64 bit mac os x build working!!!

+++ CloudsLibrary/src/VisualSystems/Histogram/vs_src/CloudsVisualSystemHistogram.cpp
@@ -313,7 +313,10 @@ void CloudsVisualSystemHistogram::selfDraw()
 	glFogi(GL_FOG_COORD_SRC, GL_FRAGMENT_DEPTH);
 	glFogi(GL_FOG_MODE, GL_EXP);
     
-    GLfloat fogColor[4] = { bgColor.r/255., bgColor.g/255., bgColor.b/255., 1.0 };
+    GLfloat fogColor[4] = {
+		(GLfloat)(bgColor.r/255.),
+		(GLfloat)(bgColor.g/255.),
+		(GLfloat)(bgColor.b/255.), (GLfloat)(1.0) };
     glFogfv(GL_FOG_COLOR, fogColor);
     
 	glFogf(GL_FOG_DENSITY, powf(fogDensity, 4));
